Migrate to Scala3 (#132)
This PR exists to update our BE repo from Scala2 to Scala3

Key Changes:
- No longer using the keyword implicit and have been replaced with given
and using appropriately
- Fix JSON and BSON codec usage on repository classes after Scala 3
upgrade
- General refactoring

+++ app/uk/gov/hmrc/pillar2/config/AppConfig.scala
@@ -26,7 +26,7 @@ class AppConfig @Inject() (val config: Configuration, servicesConfig: ServicesCo
 
   val appName: String = config.get[String]("appName")
 
-  val defaultDataExpireInSeconds: Int = config.get[Int]("defaultDataExpireInSeconds")
+  val defaultDataExpireInSeconds: Long = config.get[Long]("defaultDataExpireInSeconds")
 
   def baseUrl(serviceName: String): String =
     s"${servicesConfig.baseUrl(serviceName)}${servicesConfig.getString(s"microservice.services.$serviceName.context")}"
